+++ lib/parsers/widget_parser.dart
@@ -1,5 +1,7 @@
 import 'package:flutter/widgets.dart';
 
+import '../json_to_form_with_theme.dart';
+
 abstract class WidgetParser {
   Widget getWidget();
   late String id;